Rename deforamble_torus example to deformable
