+++ examples/multibody/deformable/BUILD.bazel
