+++ examples/multibody/deformable/parallel_gripper_controller.h
