+++ examples/multibody/deformable/point_source_force_field.h
