+++ examples/multibody/deformable/simple_gripper.sdf
