+++ examples/multibody/deformable/torus.vtk
